First draft of the Sierra GUI implementation
Created the `g.gd` script to test an approach for accessing the commands
defined for the graphic interface.
Moved the Toolbar folder to the popochiu template folder since that
component will be part of that GUI only.
Added a function to "replace" the default cursor. This still needs
improvements or a complete refactor.