9 verb inventory scroll working
It takes into account the size of each slot, the size of the gaps
between rows and the number of rows.
This inventory panel adds items to the i_inventory based on what the
scene tree has.
NEXT >> Make this as a component that can work in a popup (i.e. in the
Sierra inventory popup).