Updated command handling
Both in PopochiuClickable and PopochiuInventoryItem. They will call
`_on_click` if not custom method is present in the object. Method
fallback should be manually called with `E.command_fallback()`.
Added script_name property to PopochiuCommands class. This can be used
to know which set of commands is the current one.