Fix and update in Setup popup
Now it shows a description of the selected template so devs can know a
bit more about each one.
If the selected template is the same that is currently active after
closing the Setup popup, no action is performed.
Add Popsy as a character to test a bug reported related to z-index.
Add GUI cursor to BASS cursors.