Create script copies when selecting GUI template
A copy of the template scripts for the graphic interface and the
commands linked to it is created and placed in
res://popochiu/graphic_interface.
This makes sure devs can play with those scripts instead of modifying
those in the addons folder.

fea: Create a commands template script for the Empty template.